Add website digest synchronization  (#5362)
* Add production branch to PR review workflow

* Rename discord_digest.yml to weekly_digest.yml

Add website digest sync along with discord digest post

* Add branch filters 

Only include merged PRs from main, master and production branches within the 7 days period

* Add website news generator script

This script fetches merged pull requests from a GitHub repository, generates concise news entries for a website based on those changes, and updates a news file in the repository with the generated content.

* Include the digest as PRs description 

This will include the actual news in the description of the PR so it will be easy to review.